Added User Account Confirmation
The Frontend and the backend are now working succesfully on the
localhost. The same is not true of the published branch at the moment,
but I will fix that later. A user can register and once they click the
link to get their verification token it will do the processing on the
backend and send the result to the server.